refactor: move RichTextEditorSkeleton to a separate file and clean up imports
